Change headers and choices in 2020 leader application
This commit makes a number of changes to the leader application. Because
time is a concern, efforts were made the limit the extent to which these
changes would impact the underlying database to minimize potential
impact. In most cases, the new fields are tweaks to how the options are
displayed, rather than changes that would be reflected in the database.