Extend trip duration by one day
Due to Coronavirus, all trip leaders will have to arrive a day early to
finish their training. Right now, the key trip dates are pegged off of
the leader arrival i.e. the database knows that the first night camping
will take place two days after the leaders arrived. This won't be the
case for 2020.

In order to make this maximally flexible (and realtively easy to
revert), the previous commit moved the "deltas" into a constants file
that is accessed by both the trips managers and the models. In this
commit, I edited the constants and all the tests that are built off
those deltas as well. I left the test-values hardcoded for now, since
they are functionality that would break if you changed the values.