Force database to update current students' display names
Students now have the option to update their names in the Dartmouth
directory, but if they do so after having logged into the trips
database, the database will have no knowledge of it, since the directory
retrieval only happen on account creation.

This migration pings DartDM for a new display name for every student in
classes 2019-2023. A forthcoming update with perform a lookup on login,
but it was necessary to quickly update all applicants to avoid
deadnaming them.