Fix Django upgrade errors
The new Django version broke the import of a FieldDoesNotExist error and
the functionality of the __in modifier. Fixing the import was trivial;
for the latter, it appears that you can no longer use this on an
unevaluated values_list without triggering a "more than one row returned
by a subquery used as an expression" SQL error.

I fixed this error by forcing it to make the call and returning a list.
According to the Django docummentation, the benefit of a values_list is
that it can lazy-slice. In the case here, we're trying to determine if a
primary key is in the values_list, so the performance benefits from
lazy-slicing there are, from a complexity standpoint, negligible.